% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaksAlignment.R
\docType{methods}
\name{compress,peaksAlignment-method}
\alias{compress,peaksAlignment-method}
\title{Compression method for peaksAlignment object}
\usage{
\S4method{compress}{peaksAlignment}(object, verbose = TRUE, ...)
}
\arguments{
\item{object}{peaksAlignment}

\item{verbose}{logical}

\item{...}{}
}
\description{
Compression method for peaksAlignment object
}
\author{
MR
}
\keyword{internal}
