% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressiveAlignment.R
\docType{methods}
\name{compress,progressiveAlignment-method}
\alias{compress,progressiveAlignment-method}
\title{Compress method for progressiveAlignment}
\usage{
\S4method{compress}{progressiveAlignment}(object, verbose = TRUE, ...)
}
\arguments{
\item{object}{dummy}

\item{verbose}{dummy}

\item{...}{dummy}
}
\description{
Decompress method for progressiveAlignment
}
\details{
Deompress method for progressiveAlignment
}
\author{
MR
}
\keyword{internal}
