% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addXCMSPeaks.R
\name{de_Duper}
\alias{de_Duper}
\title{deDuper}
\usage{
de_Duper(object, mz_abs = 0.1, rt_abs = 2)
}
\arguments{
\item{object}{xcms object}

\item{mz_abs}{mz range}

\item{rt_abs}{rt range}
}
\value{
an object of xcms class
}
\description{
Duplicate peak removal function
}
\details{
Remove redundant peaks, in this case where there are any peaks within an
absolute m/z value of 0.2 and within 3 s for any one sample in the xcmsSet
(the largest peak is kept)
}
\author{
r
}
