% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaksAlignment.R
\docType{methods}
\name{decompress,peaksAlignment-method}
\alias{decompress,peaksAlignment-method}
\title{Decompression method for peaksAlignment object}
\usage{
\S4method{decompress}{peaksAlignment}(object, verbose = TRUE, ...)
}
\arguments{
\item{object}{peaksAlignment object}

\item{verbose}{dummy}

\item{...}{dummy}
}
\description{
Decompression method for peaksAlignment object
}
\author{
MR
}
\keyword{internal}
