% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressiveAlignment.R
\name{decompress,progressiveAlignment-method}
\alias{decompress,progressiveAlignment-method}
\title{Compress method for progressiveAlignment}
\usage{
\S4method{decompress}{progressiveAlignment}(object, verbose = TRUE, ...)
}
\arguments{
\item{object}{progressiveAlignment object}

\item{verbose}{logical}

\item{...}{dummy}
}
\description{
Decompress method for progressiveAlignment
}
\details{
Decompress method for progressiveAlignment
}
\author{
MR
}
\keyword{internal}
