% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addXCMSPeaks.R
\name{deDuper}
\alias{deDuper}
\title{deDuper}
\usage{
deDuper(object, mz.abs = 0.1, rt.abs = 2)
}
\arguments{
\item{object}{xcms object}

\item{mz.abs}{mz range}

\item{rt.abs}{rt range}
}
\value{
an object of xcms class
}
\description{
Duplicate peak removal function
}
\details{
Remove redundant peaks, in this case where there are any peaks within an
absolute m/z value of 0.2 and within 3 s for any one sample in the xcmsSet
(the largest peak is kept)
}
\author{
r
}
