% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSpectra.R
\name{distToLib}
\alias{distToLib}
\title{distToLib}
\usage{
distToLib(mspLib, outList)
}
\arguments{
\item{mspLib}{a .msp file from NIST}

\item{outList}{an object from gatherInfo()}
}
\value{
the distance matrix between the mass spec and the aligned spec
}
\description{
The function calculate the distance between each mas spec in the msp file
and the aligned mass spec from each sampe
}
\details{
Return the distance matrix
}
\author{
Riccardo Romoli
}
