\name{eitherMatrix-class}
\alias{eitherMatrix-class}

\title{The \code{eitherMatrix} class}

\description{A container to store either matrix or matrix.csc objects}

\author{Mark Robinson}

\references{
Mark D Robinson (2008).  Methods for the analysis of gas chromatography - mass spectrometry data \emph{PhD dissertation} University of Melbourne.
}

\seealso{
\code{\link{peaksAlignment}}
}

\keyword{classes}
