% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportSpectra.R
\name{exportSpectra}
\alias{exportSpectra}
\title{exportSpectra}
\usage{
exportSpectra(object, outList, spectra, normalize = TRUE)
}
\arguments{
\item{object}{an object of class "peaksDataset"}

\item{outList}{an object created using the gatherInfo() function}

\item{spectra}{numeric. The number of the mass spectra to be printed. It
correspond to the number of the peak in the plot() and the number of the
peak in the gatherInfo() list.}

\item{normalize}{logical. If the mass spectra has to be normalized to 100}
}
\value{
a .msp file
}
\description{
Write the mass spectum into a .msp file to be used in NIST search.
}
\details{
Write the mass spectum into a .msp file to be used in NIST search.
}
\author{
riccardo.romoli@unifi.com
}
