% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSpectra.R
\name{headToTailPlot}
\alias{headToTailPlot}
\title{Head to tail plot}
\usage{
headToTailPlot(specFromLib, specFromList)
}
\arguments{
\item{specFromLib}{the mass spectra obtained from the .msp file}

\item{specFromList}{the mass spectra obtained from \code{\link{gatherInfo}}}
}
\value{
the plot
}
\description{
The head-to-tail-plot for the mass spectra
}
\details{
Head-to-tail-plot to visually compare the mass spectra
}
\author{
Riccardo Romoli
}
