% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSpectra.R
\name{importSpec}
\alias{importSpec}
\title{importSpec}
\usage{
importSpec(file)
}
\arguments{
\item{file}{a .msp file from NIST search library database}
}
\value{
list conaining the mass spctra
}
\description{
Read the mass spectra from an external msp file
}
\details{
Read the mass spectra from an external file in msp format. The format is
used in NIST search library database.
}
\author{
riccardo.romoli@unifi.it
}
