% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSpectra.R
\name{matchSpec}
\alias{matchSpec}
\title{matchSpec}
\usage{
matchSpec(spec1, outList, whichSpec)
}
\arguments{
\item{spec1}{reference mass spectrum}

\item{outList}{the return of \code{\link{gatherInfo}}}

\item{whichSpec}{the entry number of outList}
}
\value{
the distance between the reference mass spectrum and the others
}
\description{
Calculate the distance between a reference mass spectrum
}
\details{
Calculate the distance between a reference mass spectrum and one from the
sample
}
\author{
Riccardo Romoli
}
