% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpectra.R
\name{plotMultipleSpectra}
\alias{plotMultipleSpectra}
\title{plotMultipleSpectra}
\usage{
plotMultipleSpectra(object, outList, spectra, fullRange = TRUE,
  normalize = TRUE, ...)
}
\arguments{
\item{object}{where to keep the mass range of the experiment}

\item{outList}{where to keep the mass spectra; both abundance than m/z}

\item{spectra}{a vector containing the index of the spectra to be plotted.
Is referred to outList}

\item{fullRange}{if TRUE uses the mass range of the whole experiment, 
otherwise uses only the mass range of each plotted spectum}

\item{normalize}{if TRUE normalize the intensity of the mass peak to 100, 
the most abundant is 100% and the other peaks are scaled 
consequetially}

\item{...}{further arguments passed to the ‘plot’ command}
}
\description{
Plot the aligned mass spectra
}
\details{
Plot the deconvoluted and aligned mass spectra collected using gatherInfo()
}
\examples{
## Rd workflow
gcmsPath <- paste(find.package("gcspikelite"), "data", sep = "/")
cdfFiles <- dir(gcmsPath,"CDF", full = TRUE)

# read data, peak detection results
pd <- peaksDataset(cdfFiles[1:4], mz = seq(50,550), rtrange = c(7.5,10.5))
pd <- addXCMSPeaks(files = cdfFiles[1:4], object = pd, peakPicking = c('mF'),
                   snthresh = 2, fwhm = 8,  step = 0.5, steps = 2, mzdiff = 0.5,
                   sleep = 0)
## multiple alignment
ma <- multipleAlignment(pd, c(1,1,2,2), wn.gap = 0.5, wn.D = 0.05, bw.gap = 0.6, 
                        bw.D = 0.2, usePeaks = TRUE, filterMin = 1, df = 50,
                        verbose = TRUE, metric = 2, type = 2)

## gather apex intensities
gip <- gatherInfo(pd, ma)
gip[[33]]
plotMultipleSpectra(object = pd, outList = gip, spectra = 33, fullRange = FALSE,
                    normalize = TRUE)
}
\author{
Riccardo Romoli (riccardo.romoli@unifi.it)
}
\keyword{gatherInfo()}
\keyword{plot()}
