% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleAlignment.R
\docType{methods}
\name{show,multipleAlignment-method}
\alias{show,multipleAlignment-method}
\title{Store the raw data and optionally, information regarding signal peaks for
a number of GCMS runs}
\usage{
\S4method{show}{multipleAlignment}(object)
}
\arguments{
\item{object}{multipleAlignment object}
}
\description{
multipleAlignment is the data structure giving the result of an alignment 
across several GCMS runs. 	Multiple alignments are done progressively.
First, all samples with the same \code{tg$Group} label with be aligned
(denoted a "within" alignment).  Second, each group will be summarized into
a pseudo-data set, essentially a spectrum and retention time for each matched
peak of the within-alignment.  Third, these "merged peaks" are aligned in the
same progressive manner, here called a "between" alignment.
}
\author{
Mark Robinson
}
