% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{corPrt}
\alias{corPrt}
\title{Retention Time Penalized Correlation}
\usage{
corPrt(d1, d2, t1, t2, D, penality = 0.2)
}
\arguments{
\item{d1}{data matrix for sample 1}

\item{d2}{data matrix for sample 2}

\item{t1}{vector of retention times for sample 1}

\item{t2}{vector of retention times for sample 2}

\item{D}{retention time window for the matching}

\item{penality}{penalization applied to the matching between two mass
spectra if \code{(t1-t2)>D}}
}
\value{
matrix of similarities
}
\description{
This function calculates the similarity of all pairs of peaks from 2
samples, using the spectra similarity and the rretention time differencies
}
\details{
Computes the Pearson carrelation between every pair of peak vectors in the
retention time window (\code{D})and returns the similarity matrix.
}
\examples{

## Not Run
require(gcspikelite)
files <- list.files(path = paste(find.package("gcspikelite"), "data",
                    sep = "/"),"CDF", full = TRUE)
data <- peaksDataset(files[1:2], mz = seq(50, 550), rtrange = c(7.5, 8.5))
## create settings object
mfp <- xcms::MatchedFilterParam(fwhm = 10, snthresh = 5)
cwt <- xcms::CentWaveParam(snthresh = 3, ppm = 3000, peakwidth = c(3, 40),
 prefilter = c(3, 100), fitgauss = FALSE, integrate = 2, noise = 0,
 extendLengthMSW = TRUE, mzCenterFun = "wMean")
data <- addXCMSPeaks(files[1:2], data, settings = mfp, minintens = 100,
 multipleMatchedFilter = FALSE, multipleMatchedFilterParam =
 list(fwhm = c(5, 10, 20), rt_abs = 3, mz_abs = 0.1))
data
## review peak picking
plotChrom(data, rtrange=c(7.5, 10.5), runs=c(1:2))

r <- corPrt(data@peaksdata[[1]], data@peaksdata[[2]],
           data@peaksrt[[1]], data@peaksrt[[2]], D = 50, penality = 0.2)
## End (Not Run)

}
\seealso{
\code{\link{peaksAlignment}}
}
\author{
Riccardo Romoli
}
\keyword{manip}
