% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp.R
\name{dynRT}
\alias{dynRT}
\title{dynRT}
\usage{
dynRT(S)
}
\arguments{
\item{S}{similarity matrix}
}
\value{
list containing the matched peaks between the two
    chromatograms. The number represent position of the spectra in
    the S matrix
}
\description{
Dynamic Retention Time Based Alignment algorithm, given a similarity matrix
}
\details{
This function align two chromatograms finding the maximum
    similarity among the mass spectra
}
\examples{
require(gcspikelite)
gcmsPath <- paste(find.package("gcspikelite"), "data", sep="/")
cdfFiles <- dir(gcmsPath,"CDF", full=TRUE)
## read data, peak detection results
pd <- peaksDataset(cdfFiles[1:3], mz=seq(50,550),
    rtrange=c(7.5,10.5))
pd <- addXCMSPeaks(files=cdfFiles[1:3], object=pd,
    peakPicking=c('mF'),snthresh=3, fwhm=10,  step=0.1, steps=2,
    mzdiff=0.5, sleep=0)
## review peak picking
plot(pd, rtrange=c(7.5, 10.5), runs=c(1:3))
## similarity
r <- ndpRT(pd@peaksdata[[1]], pd@peaksdata[[2]], pd@peaksrt[[1]],
    pd@peaksrt[[2]], D=50)
## dynamic retention time based alignment algorithm
v <- dynRT(S=r)
}
\author{
riccardo.romoli@unifi.it
}
