% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportSpectra.R
\name{exportSpectra}
\alias{exportSpectra}
\title{exportSpectra}
\usage{
exportSpectra(object, sample, spectraID, normalize = TRUE)
}
\arguments{
\item{object}{an object of class "peaksDataset" where to keep the
mass spectra; both abundance (y) than m/z (x)}

\item{sample}{character, the sample from were to plot the mass spectra}

\item{spectraID}{numerical, a vector containing the index of the spectra to be
plotted.}

\item{normalize}{logical, if TRUE normalize the intensity of the
mass peak to 100, the most abundant is 100% and the other peaks
are scaled consequetially}
}
\value{
a .msp file ready to be read using NIST search
}
\description{
Write the deconvoluted mass spectra to an external file
}
\details{
Write a .msp file of the deconvoluted mass spectra. Usfull to try
to identify the unknown spectra using NIST Search.
}
\author{
riccardo.romoli@unifi.it
}
