% Generated by roxygen2: do not edit by hand
% Please edit documentation in RtmpN0olRm/file2ef76f9acdc3
\name{importSpec}
\alias{importSpec}
\title{importSpec}
\usage{
importSpec(file)
}
\arguments{
  \item{file}{}
}
\value{
list conaining the mass spctra
}
\description{
Read the mass spectra from an external msp file
}
\details{
Read the mass spectra from an external file in msp format.
The format is used in NIST search library database.
}
\author{
riccardo.romoli@unifi.it
