% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaksAlignment.R
\docType{methods}
\name{plotAlignment,peaksAlignment-method}
\alias{plotAlignment,peaksAlignment-method}
\title{plotAlignment}
\usage{
\S4method{plotAlignment}{peaksAlignment}(object, xlab = "Peaks - run 1",
  ylab = "Peaks - run 2", plotMatches = TRUE, matchPch = 19,
  matchLwd = 3, matchCex = 0.5, matchCol = "black",
  col = colorpanel(50, "white", "green", "navyblue"), breaks = seq(0,
  1, length = 51), ...)
}
\arguments{
\item{object}{a \code{clusterAlignment} object}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{plotMatches}{logical, whether to plot matches}

\item{matchPch}{match plotting character}

\item{matchLwd}{match line width}

\item{matchCex}{match character expansion factor}

\item{matchCol}{match colour}

\item{col}{vector of colours for colourscale}

\item{breaks}{vector of breaks for colourscale}

\item{...}{further arguments passed to \code{image}}
}
\value{
plot an object of class \code{\linkS4class{peaksAlignment}}
}
\description{
Plotting functions for GCMS data objects
}
\details{
Plot an object of \code{\linkS4class{peaksAlignment}}

The similarity matrix is plotted and optionally, the set of matching peaks.
\code{clusterAlignment} objects are just a collection of all pairwise
\code{peakAlignment} objects.
}
\examples{

require(gcspikelite)

## paths and files
gcmsPath <- paste(find.package("gcspikelite"), "data", sep="/")
cdfFiles <- dir(gcmsPath, "CDF", full=TRUE)
eluFiles <- dir(gcmsPath, "ELU", full=TRUE)

## read data
pd <- peaksDataset(cdfFiles[1:3], mz=seq(50,550), rtrange=c(7.5,8.5))
pd <- addXCMSPeaks(files=cdfFiles[1:3], object=pd, peakPicking=c('mF'),
                   snthresh=3, fwhm=10,  step=0.1, steps=2, mzdiff=0.5)

## image plot
plotChrom(pd, rtrange=c(7.5,8.5), plotPeaks=TRUE, plotPeakLabels=TRUE)

## align two chromatogram
pA <- peaksAlignment(pd@peaksdata[[1]], pd@peaksdata[[2]],
                     pd@peaksrt[[1]], pd@peaksrt[[2]], D = 50,
                     compress = FALSE, type = 1, metric = 1,
                     gap = 0.5)
plotAlignment(pA)

}
\references{
Mark D Robinson (2008).  Methods for the analysis of gas
chromatography - mass spectrometry data \emph{PhD dissertation} University
of Melbourne.
}
\seealso{
\code{\link{peaksAlignment}} \code{\link{plotAlignment}}
}
\author{
Mark Robinson
}
\keyword{classes}
