% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterAlignment.R
\docType{methods}
\name{plotClustAlignment,clusterAlignment-method}
\alias{plotClustAlignment,clusterAlignment-method}
\title{plotClustAlignment}
\usage{
\S4method{plotClustAlignment}{clusterAlignment}(object, alignment = 1,
  ...)
}
\arguments{
\item{object}{\code{clusterAlignment} object.}

\item{alignment}{the set of alignments to plot}

\item{...}{further arguments passed to \code{image}. See
also \code{plotAlignment}}
}
\value{
plot the pairwise alignment
}
\description{
Plotting functions for GCMS data objects
}
\details{
For \code{clusterAlignment} objects, the similarity matrix is plotted and
optionally, the set of matching peaks.  \code{clusterAlignment} objects are
just a collection of all pairwise \code{peakAlignment} objects.
}
\examples{

require(gcspikelite)

## paths and files
gcmsPath <- paste(find.package("gcspikelite"), "data", sep="/")
cdfFiles <- dir(gcmsPath, "CDF", full=TRUE)
eluFiles <- dir(gcmsPath, "ELU", full=TRUE)

## read data
pd <- peaksDataset(cdfFiles[1:3], mz=seq(50,550), rtrange=c(7.5,8.5))
pd <- addXCMSPeaks(files=cdfFiles[1:3], object=pd, peakPicking=c('mF'),
                   snthresh=3, fwhm=10,  step=0.1, steps=2, mzdiff=0.5)
ca <- clusterAlignment(pd, metric = 1, D = 50, type = 1, gap = 0.5)
plotClustAlignment(ca, run = 1)
plotClustAlignment(ca, run = 2)
plotClustAlignment(ca, run = 3) 

}
\references{
Mark D Robinson (2008).  Methods for the analysis of gas
chromatography - mass spectrometry data \emph{PhD dissertation} University
of Melbourne.
}
\seealso{
\code{\link{plotAlignment}}
}
\author{
Mark Robinson
}
\keyword{classes}
