% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retFatMatrix.R
\name{retFatMatrix}
\alias{retFatMatrix}
\title{retFatMatrix}
\usage{
retFatMatrix(object, data, minFilter = round(length(object@files)/3 * 2))
}
\arguments{
\item{object}{peakDataset object}

\item{data}{a gatherInfo() object}

\item{minFilter}{the minimum number for a feature to be returned in the data
matrix. Default is 2/3 of the samples}
}
\value{
A fat data matrix containing the area of the deconvoluted and
aligned peaks. The row are the samples while the column represent the
different peaks
}
\description{
Build a fat data matrix
}
\details{
This function allows to extract the data from an object created using
\code{gatherInfo} and build a data matrix using the area of the deconvoluted
and aligned peaks. The row are the samples while the column represent the
different peaks.
}
\examples{

require(gcspikelite)
# paths and files
gcmsPath <- paste(find.package("gcspikelite"), "data", sep = "/")
cdfFiles <- dir(gcmsPath,"CDF",full=TRUE)
# read data, peak detection results
pd <- peaksDataset(cdfFiles[1:2], mz=seq(50,550),
                   rtrange=c(7.5,8.5))
pd <- addXCMSPeaks(files=cdfFiles[1:2], object=pd,
                   peakPicking=c('mF'), snthresh=3, fwhm=4,
                   step=1, steps=2, mzdiff=0.5)
ma <- multipleAlignment(pd = pd, group = c(1,1),
                        filterMin = 1, metric = 2, type = 2)
outList <- gatherInfo(pd, ma)
mtxD <- retFatMatrix(object = pd, data = outList, minFilter = 1)

}
\seealso{
\code{\link{gatherInfo}}
}
\author{
Riccardo Romoli \email{riccardo.romoli@unifi.it}
}
