% Generated by roxygen2: do not edit by hand
% Please edit documentation in Rtmpji4S9w/file274b55601379
\name{retFatMatrix}
\alias{retFatMatrix}
\title{retFatMatrix}
\usage{
retFatMatrix(data)
}
\arguments{
  \item{data}{the list obtained from gatherInfo()}
}
\value{
A fat data matrix containing the area of the deconvoluted
and aligned peaks. The row are the samples while the column
represent the different peaks.
}
\description{
Build a fat data matrix
}
\details{
This function allows to extract the data from an object
created using \code{gatherInfo} and build a data matrix
using the area of the deconvoluted and aligned peaks. The
row are the samples while the column represent the different
peaks.
}
\examples{
require(gcspikelite)
# paths and files
gcmsPath <- paste(find.package("gcspikelite"), "data", sep="/")
cdfFiles <- dir(gcmsPath,"CDF",full=TRUE)
# read data, peak detection results
pd <- peaksDataset(cdfFiles[1:2], mz=seq(50,550),
                   rtrange=c(7.5,8.5))
pd <- addXCMSPeaks(files=cdfFiles[1:2], object=pd,
                   peakPicking=c('mF'), snthresh=3, fwhm=4,
                   step=1, steps=2, mzdiff=0.5)
mp <- correlationAlignment(object=pd, thr=0.85, D=20,
                           penality=0.2, normalize=TRUE,
                           minFilter=1) 
outList <- gatherInfo(pd, mp)
mtxD <- retFatMatrix(data=outList)
}
\author{
Riccardo Romoli \email{riccardo.romoli@unifi.it}
}
\seealso{
\code{\link{gatherInfo}}
}
